refactor: Tái cấu trúc và cải thiện mã nguồn
Tái cấu trúc dự án bằng cách chia nhỏ main.py thành các module riêng biệt:
- app.py: Chứa logic UI và widget chính.
- logic.py: Chứa logic game và quản lý trạng thái.
- config.py: Chứa các hằng số và định nghĩa sprite.
- input.py: Xử lý lắng nghe bàn phím toàn hệ thống.

Mục tiêu là cải thiện cấu trúc mã nguồn, tăng khả năng đọc và dễ bảo trì.

Closes #2
